#86 refactor : Extensions 함수 파일명 수정, 파일분리
