+++ Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/DataExtensions.kt
@@ -1,0 +1,13 @@
+package com.avengers.nibobnebob.presentation.ui
+
+import java.text.SimpleDateFormat
+import java.util.Date
+import java.util.Locale
+import java.util.TimeZone
+
+internal fun Long.toDateString(): String {
+    val dateFormat = SimpleDateFormat("yyyy/MM/dd", Locale.getDefault())
+    dateFormat.timeZone = TimeZone.getTimeZone("UTC")
+    val date = Date(this)
+    return dateFormat.format(date)
+}
