+++ Aos/app/src/main/java/com/avengers/nibobnebob/presentation/ui/UiExtensions.kt
@@ -1,10 +1,10 @@
-package com.avengers.nibobnebob.presentation.util
+package com.avengers.nibobnebob.presentation.ui
 
 import android.content.Context
 import android.view.View
 import android.view.inputmethod.InputMethodManager
 
-fun Context.hideKeyboard(view: View) {
+internal fun Context.hideKeyboard(view: View) {
     val imm = getSystemService(Context.INPUT_METHOD_SERVICE) as InputMethodManager
     imm.hideSoftInputFromWindow(view.windowToken, 0)
 }
